make SootMethod as interface
