+++ sootup.java.core/src/test/java/sootup/java/core/model/SootMethodTest.java
@@ -101,7 +101,7 @@ public class SootMethodTest {
   public void constructorTest() {
     IdentifierFactory idf = JavaIdentifierFactory.getInstance();
 
-    SootMethod mockedConstructor = spy(new SootMethod.SootMethodBuilder().build());
+    SootMethod mockedConstructor = spy(new JavaSootMethod.SootMethodBuilder().build());
     MethodSignature mockedSignature = mock(MethodSignature.class);
     MethodSubSignature mockedSubSignature = mock(MethodSubSignature.class);
